refact: type separated from useProductInfo hook
